requirement.txt
amqp==5.0.9
asgiref==3.4.1
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
backports.zoneinfo==0.2.1
bidict==0.21.4
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==3.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
eventlet==0.33.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-tools==1.43.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.3
pika==1.2.0
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-engineio==4.3.1
python-socketio==5.5.1
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
vine==5.0.0
zope.interface==5.4.0
